requirements.txt


python-dotenv
PyJWT
requests
langchain-community
langchain-openai

psutil
jieba
opencc-python-reimplemented
docx2txt
ebooklib
beautifulsoup4
lxml
chromadb
SQLAlchemy
pydantic

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0

langchain>=0.0.335
psycopg2-binary

jinja2
python-multipart

starlette 
